+++ hal/linux/spi/libtropic_port_linux_spi.h
@@ -11,7 +11,7 @@
 
 #include <linux/gpio.h>
 
-#include "libtropic_port.h"
+#define LT_SPI_PATH_MAX_LEN 256
 
 #ifdef __cplusplus
 extern "C" {
@@ -27,9 +27,9 @@ typedef struct lt_dev_linux_spi_t {
     /** @public @brief SPI speed in Hz. */
     int spi_speed;
     /** @public @brief Path to the SPI device. */
-    char spi_dev[LT_DEVICE_PATH_MAX_LEN];
+    char spi_dev[LT_SPI_PATH_MAX_LEN];
     /** @public @brief Path to the GPIO device. */
-    char gpio_dev[LT_DEVICE_PATH_MAX_LEN];
+    char gpio_dev[LT_SPI_PATH_MAX_LEN];
     /** @public @brief Number of the GPIO pin to map chip select to. */
     int gpio_cs_num;
 #if LT_USE_INT_PIN